Name Patterns for Gender
This python file shows the name patterns for gender based on the last
name. You can change this to see the patterns based on suffixes, number
of letters, etc.
This is tutorial is present in the book ‘Natural Language Processing
with Python’ by Steven Bird, Ewan Klein and Edward Loper.